Work-in-progress FT800 interface.
This is an adapter from the parallel system buses to the SPI bus used by
the FT800, turning reads and writes into the appropriate SPI sequences.
At least, it will be once it's finished. Currently it's broken, as
illustrated by the tests.

The idea is to clock this much faster than the CPU (although still under
30MHz since that is the highest rate supported by the FT800) so that
the read/write delays are minimized, but accessing the graphics chip
memory will still cause some wait states when the CPU is running at a
reasonable speed, unfortunately.