Support the Bus Error signal in fake68k reads.
Bus Error makes the read cycle two clocks longer and allows the memory
controller to halt processing until the bus error signal is un-asserted.

The exception processing that would normally follow a bus error is not
simulated, since that just entails a pair of standard read cycles to
obtain the exception handler address.