Just enough gdb server to connect and continue
This is a shaky implementation of the gdb remote protocol that's just
enough for gdb to connect and continue execution at the start of kernel
initialization.

Once we've continued there's no way to get back into the debugger, so this
isn't particularly useful but is a good starting point.

As a side-effect of getting this working, it's also possible to run
"info registers" to dump out the initial register values, though of course
their states at boot are well-known and so little value is derived from
asking gdb about them.