Make breakpoints behave correctly in the emulator
Stock Musashi doesn't allow the instruction hook to abort processing the
instruction, meaning that previously our debugger stub would execute one
instruction too many when encountering a breakpoint.

Here we've lightly patched Musashi to add a special case where it checks
if the instruction hook stole the CPU's timeslice, and if so it will
terminate the loop immediately without executing the instruction.