Generate a tick interrupt every 200ms in the emulator
This will be our scheduler's timeslice signal, allowing it to preempt a
user process.

Since we don't yet have the real interrupt controller designed, and thus
can't emulate that design, for now we just assume that the timeslice
signal will be priority 1 and that it will vector to user interrupt vector
64, which gives us something to hook the OS up to for now, pending some
final definition of how this is going to look.