When tracing emulated instructions, don't print on break
Now that we break before executing an instruction at a breakpoint, we
shouldn't print out that instruction's disassembly in instruction trace
mode, since that's misleading.

We still print it out in step mode, since we will execute this one
instruction before we end our timeslice and return control to the
debugger.