emulator: use a more accurate cycle-based emulation style
Previously this was handled in a rather lazy way because we'd not yet
figured out how the display clock and CPU clock would interact, what
other interrupts would be present, etc.

Now that we're starting to solidify the specification of the machine, it's
better to match a little better the "real" timing behavior, so we can
start to introduce timer interrupts properly, etc.