video: Start of video controller; just pixel clock for the moment
Here we use a PLL block to multiply the 12MHz clock on the ICEbreaker
board I'm currently using for development to create a 74.25MHz pixel clock
for 720p video output.

However, we're not yet generating any other signals, so for the moment
this is useless and just puts a sine wave on one of our pins.