video: Further progress on the mouse pointer images
Something's still not quite right here, but at least the image data seems
to now be in a reasonable shape and the X,Y coordinates are being selected
propertly.

The decoding is still "off" somehow, selecting the wrong color indices for
pixels and behaving differently in the top and bottom half of the sprite.